feat(verification): extract verification logic into separate module
refactor(auth): move verification-related code to verification module
style: clean up trailing whitespace in multiple files
