+++ libs/integrations/aliyun/aliyun-sms.service.ts
@@ -3,7 +3,7 @@ import Dysmsapi20170525, * as $Dysmsapi20170525 from '@alicloud/dysmsapi20170525
 import * as $OpenApi from '@alicloud/openapi-client';
 import * as $Util from '@alicloud/tea-util';
 import Credential from '@alicloud/credentials';
-import { CodeType } from '@/auth/dto/auth.dto';
+import { CodeType } from '@/verification/enums/code-type.enum';
 
 @Injectable()
 export class AliyunSmsService {
