+++ libs/integrations/email/email.module.ts
@@ -8,4 +8,3 @@ import { MailerService } from './mailer.service';
   exports: [MailerService],
 })
 export class MailerModule {}
-
