+++ libs/integrations/email/index.ts
@@ -1,3 +1,2 @@
 export * from './email.module';
 export * from './mailer.service';
-
