+++ libs/security/decorators/user.decorator.ts
@@ -19,4 +19,3 @@ export const User = createParamDecorator(
     return data ? user?.[data] : user;
   },
 );
-
