+++ src/auth/auth.module.ts
@@ -4,12 +4,8 @@ import { PassportModule } from '@nestjs/passport';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { AuthController } from './auth.controller';
 import { AuthService } from './services/auth.service';
-import { VerificationService } from './services/verification.service';
-import { AliyunModule } from '@libs/integrations/aliyun/aliyun.module';
-import { AliyunSmsService } from '@libs/integrations/aliyun/aliyun-sms.service';
 import { JwtStrategy, JwtAuthGuard } from '@libs/security';
 import { PrismaService } from '@/prisma.service';
-import { VerificationRepository } from './repositories/verification.repository';
 import { EmailModule } from '@/email/email.module';
 import { AuthRepository } from './repositories/auth.repository';
 
@@ -26,18 +22,14 @@ import { AuthRepository } from './repositories/auth.repository';
       }),
       inject: [ConfigService],
     }),
-    AliyunModule,
     EmailModule,
   ],
   controllers: [AuthController],
   providers: [
     AuthService,
-    VerificationService,
-    AliyunSmsService,
     JwtStrategy,
     JwtAuthGuard,
     PrismaService,
-    VerificationRepository,
     AuthRepository,
   ],
   exports: [AuthService, JwtAuthGuard, JwtStrategy],
