+++ src/auth/decorators/api-docs.decorator.ts
@@ -252,177 +252,6 @@ export function ApiLoginDocs() {
   );
 }
 
-// 发送验证码接口文档装饰器
-export function ApiSendCodeDocs() {
-  return applyDecorators(
-    ApiOperation({
-      summary: '发送验证码',
-      description:
-        '向指定的邮箱或手机号发送验证码。支持注册、登录、重置密码等场景。发送频率限制：同一目标60秒内只能发送一次。验证码有效期为5分钟。',
-      tags: ['Auth'],
-    }),
-    ApiConsumes('application/json'),
-    ApiProduces('application/json'),
-    ApiResponse({
-      status: 200,
-      description: '验证码发送成功，请查收邮箱或短信',
-      schema: {
-        type: 'object',
-        properties: {
-          success: {
-            type: 'boolean',
-            description: '发送是否成功',
-            example: true,
-          },
-          message: {
-            type: 'string',
-            description: '发送结果消息',
-            example: '验证码已发送，请查收',
-          },
-        },
-        example: {
-          success: true,
-          message: '验证码已发送到您的邮箱，请查收',
-        },
-      },
-    }),
-    ApiResponse({
-      status: 400,
-      description: '请求参数错误，邮箱或手机号格式不正确',
-      schema: {
-        type: 'object',
-        properties: {
-          success: { type: 'boolean', example: false },
-          message: { type: 'string', example: '邮箱格式不正确' },
-          statusCode: { type: 'number', example: 400 },
-          error: { type: 'string', example: 'Bad Request' },
-        },
-        example: {
-          success: false,
-          message: '邮箱格式不正确',
-          statusCode: 400,
-          error: 'Bad Request',
-        },
-      },
-    }),
-    ApiResponse({
-      status: 429,
-      description: '发送过于频繁，请稍后再试',
-      schema: {
-        type: 'object',
-        properties: {
-          success: { type: 'boolean', example: false },
-          message: { type: 'string', example: '发送过于频繁，请60秒后再试' },
-          statusCode: { type: 'number', example: 429 },
-          error: { type: 'string', example: 'Too Many Requests' },
-        },
-        example: {
-          success: false,
-          message: '发送过于频繁，请60秒后再试',
-          statusCode: 429,
-          error: 'Too Many Requests',
-        },
-      },
-    }),
-    ApiResponse({
-      status: 500,
-      description: '服务器内部错误，验证码发送失败',
-      schema: {
-        type: 'object',
-        properties: {
-          success: { type: 'boolean', example: false },
-          message: { type: 'string', example: '验证码发送失败，请稍后重试' },
-          statusCode: { type: 'number', example: 500 },
-          error: { type: 'string', example: 'Internal Server Error' },
-        },
-        example: {
-          success: false,
-          message: '验证码发送失败，请稍后重试',
-          statusCode: 500,
-          error: 'Internal Server Error',
-        },
-      },
-    }),
-    ApiHeader({
-      name: 'Content-Type',
-      description: '请求内容类型',
-      required: true,
-      schema: {
-        type: 'string',
-        default: 'application/json',
-      },
-    }),
-    ApiHeader({
-      name: 'User-Agent',
-      description: '用户代理信息',
-      required: false,
-      schema: {
-        type: 'string',
-        example: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
-      },
-    }),
-  );
-}
-
-// 验证验证码接口文档装饰器
-export function ApiVerifyCodeDocs() {
-  return applyDecorators(
-    ApiOperation({
-      summary: '验证验证码',
-      description:
-        '验证邮箱或手机号收到的验证码是否正确。用于注册、登录、重置密码等场景的验证码校验。',
-      tags: ['Auth'],
-    }),
-    ApiConsumes('application/json'),
-    ApiProduces('application/json'),
-    ApiResponse({
-      status: 200,
-      description: '验证码验证结果',
-      schema: {
-        type: 'object',
-        properties: {
-          valid: { type: 'boolean', description: '验证码是否有效' },
-          message: { type: 'string', description: '验证结果消息' },
-        },
-        example: {
-          valid: true,
-          message: '验证码验证成功',
-        },
-      },
-    }),
-    ApiResponse({
-      status: 400,
-      description: '请求参数错误',
-      schema: {
-        example: {
-          statusCode: 400,
-          message: ['验证码至少4位', '目标不能为空'],
-          error: 'Bad Request',
-        },
-      },
-    }),
-    ApiResponse({
-      status: 422,
-      description: '验证码错误或已过期',
-      schema: {
-        example: {
-          valid: false,
-          message: '验证码错误或已过期',
-        },
-      },
-    }),
-    ApiHeader({
-      name: 'Content-Type',
-      description: '请求内容类型',
-      required: true,
-      schema: {
-        type: 'string',
-        default: 'application/json',
-      },
-    }),
-  );
-}
-
 // 重置密码接口文档装饰器
 export function ApiResetPasswordDocs() {
   return applyDecorators(
