+++ src/auth/dto/login.dto.ts
@@ -37,4 +37,3 @@ export class LoginDto {
   @IsString()
   code?: string;
 }
-
