+++ src/auth/dto/reset-password.dto.ts
@@ -19,4 +19,3 @@ export class ResetPasswordDto {
   })
   newPassword: string;
 }
-
