+++ src/auth/dto/update-profile.dto.ts
@@ -1,5 +1,11 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsEmail, IsOptional, IsString, MinLength, Matches } from 'class-validator';
+import {
+  IsEmail,
+  IsOptional,
+  IsString,
+  MinLength,
+  Matches,
+} from 'class-validator';
 
 export class UpdateProfileDto {
   @ApiProperty({ required: false, description: '用户名' })
@@ -39,4 +45,3 @@ export class UpdateProfileDto {
   @IsString()
   bio?: string;
 }
-
