+++ src/verification/enums/code-type.enum.ts
@@ -3,4 +3,3 @@ export enum CodeType {
   LOGIN = 'login',
   RESET_PASSWORD = 'reset_password',
 }
-
